Add files via upload
Includes code to write the values to be provided to motor  for a defined function in the form of y=f(x). Lots of parameter values are not saved. All arduino code is commented out, but can easily be re-added upon the wish of the operator.

To be defined- y=f(x), limits of motor while covering this path in complete_loop, and angle offsets of limbs, measured before running.

Includes code for random leg motion